requirements.txt
backports.entry-points-selectable==1.1.0
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.9
distlib==0.3.2
filelock==3.0.12
google-api-core==2.2.2
google-api-python-client==2.32.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.53.0
httplib2==0.20.2
idna==3.3
llist==0.7.1
oauthlib==3.1.1
# Editable install with no version control (pipenv==2021.5.29)
-e /Users/mateusz/.pyenv/versions/3.8.6/lib/python3.8/site-packages
platformdirs==2.0.2
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.6
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
six==1.16.0
uritemplate==4.1.1
urllib3==1.26.7
virtualenv==20.6.0
virtualenv-clone==0.5.5
